titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/shared/payer-bank/data-access-payer-bank/src/lib/+state/payer-bank.selector.ts
import { createSelector } from '@ngrx/store';
import { selectPayerBankState } from './payer-bank.reducer';

const selectError = createSelector(selectPayerBankState, state => state.error);
const selectBankDetails = createSelector(
  selectPayerBankState,
  state => state.bankDetails
);

const selectIsLoading = createSelector(
  selectPayerBankState,
  state => state.isLoading
);
const selectTotalElements = createSelector(
  selectPayerBankState,
  state => state.totalElements
);
const selectFileResponse = createSelector(
  selectPayerBankState,
  state => state.fileResponse
);
const selectErrorLog = createSelector(
  selectPayerBankState,
  state => state.errorLog
);
export const PayerBankSelectors = {
  selectError,
  selectBankDetails,
  selectIsLoading,
  selectTotalElements,
  selectFileResponse,
  selectErrorLog
};
